reorganise main, configs and models
