+++ examples/bandgap_benchmark/configs/cartnet.yaml
