+++ examples/bandgap_benchmark/configs/cgcnn.yaml
@@ -10,7 +10,7 @@ MODEL:
 
 SOLVER:
   LR: 0.01
-  EPOCHS: 100
+  EPOCHS: 1
   NUM_RUNS: 1
 
 LOGGING:
