--- examples/bandgap_benchmark/configs/finetune/leftnet/leftnet_z.yaml
@@ -1,24 +1,0 @@
-DATASET:
- TRAIN: 'data/fine_tune/train_data.json'
- VAL: 'data/fine_tune/test_data.json'
-
-MODEL:
-  NAME: 'leftnet'
-  PRETRAINED_MODEL_PATH: 'saved_models/leftnet/z/epoch-epoch=159.ckpt'
-  INIT_FILE: 'cif_file/atom_init.json'
-  CIF_FOLDER: "cif_file"
-
-
-SOLVER:
-  LR: 0.001
-  NUM_FOLDS: 10
-  EPOCHS: 50
-  NUM_RUNS: 2
-
-LEFTNET:
-  ENCODING: 'z'
-  LAYER_FREEZE: 'none'
-
-LOGGING:
-  LOG_DIR: "saved_models"
-  LOG_DIR_NAME: "leftnet_z"
