+++ examples/bandgap_benchmark/configs/leftnet_prop.yaml
