+++ examples/bandgap_benchmark/configs/leftnet_z.yaml
