+++ examples/bandgap_benchmark/configs/svr.yaml
