+++ kale/loaddata/materials_datasets.py
@@ -10,6 +10,12 @@ from torch.utils.data import Dataset
 from kale.evaluate.metrics import GaussianDistance
 from kale.prepdata.materials_features import AtomCustomJSONInitializer
 
+import json
+from typing import Optional, Tuple
+
+import pandas as pd
+from sklearn.model_selection import train_test_split
+from torch.utils.data import DataLoader
 
 class CIFData(Dataset):
     """
@@ -26,9 +32,9 @@ class CIFData(Dataset):
         step (float, optional): Step size for Gaussian distance calculation. Default is 0.2.
     """
 
-    def __init__(self, mpids_bg, cif_folder, init_file, max_nbrs, radius, randomize, dmin=0, step=0.2):
+    def __init__(self, target_path, cif_folder, init_file, max_nbrs, radius, randomize, target_key="bg", dmin=0, step=0.2):
         self.max_num_nbr, self.radius = max_nbrs, radius
-
+        mpids_bg = self.load_data(target_path, target_key)
         if randomize:
             self.mpids_bg_dataset = (
                 mpids_bg.sample(frac=1).reset_index(drop=True).values
@@ -81,6 +87,22 @@ class CIFData(Dataset):
             batch_idx=torch.cat(batch_atom_indices),
             batch_size=len(batch_cif_ids),
         )
+    
+    def load_data(self, target_path, target_key) -> pd.DataFrame:
+        """
+        Load and return the dataset as a pandas DataFrame.
+
+        Args:
+            target_path (str): Path to the JSON file containing the dataset.
+            target_key (str): The key in the JSON file that corresponds to the target property.
+        """
+        with open(target_path, "r") as f:
+            data_json = json.load(f)
+            data_df = pd.DataFrame.from_dict(data_json, orient="index").reset_index()
+            data_df.rename(columns={"index": "mpids", 0: target_key}, inplace=True)
+
+        return data_df
+
 
     def __len__(self):
         return len(self.mpids_bg_dataset)
@@ -143,3 +165,6 @@ class CIFDataItem:
             "target": self.target,
             "cif_id": self.cif_id,
         }
+
+
+
